[WIP] checkpoint before fixing React Hook incompatibility
