+++ peak-app/package.json
@@ -116,13 +116,14 @@
     "@types/slate-react": "^0.22.9",
     "@types/styled-components": "^5.1.0",
     "@types/uuid": "^7.0.3",
-    "@udecode/slate-plugins": "v1.0.0-alpha.19",
+    "@udecode/slate-plugins": "^1.0.0-alpha.25",
     "ace-builds": "^1.4.11",
     "animate.css": "^4.1.1",
     "antd": "^4.4.1",
     "axios": "^0.19.2",
     "axios-hooks": "^1.9.0",
     "classnames": "^2.2.6",
+    "component-library": "link:../component-library",
     "copy-webpack-plugin": "^6.2.1",
     "electron-deeplink": "^1.0.4",
     "electron-is-dev": "^1.2.0",
@@ -153,10 +154,10 @@
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0",
     "sass": "^1.32.12",
-    "slate": "^0.59.0",
-    "slate-history": "^0.59.0",
-    "slate-hyperscript": "^0.59.0",
-    "slate-react": "^0.60.11",
+    "slate": "^0.63.0",
+    "slate-history": "^0.62.0",
+    "slate-hyperscript": "^0.62.0",
+    "slate-react": "^0.63.0",
     "styled-components": "^5.1.1",
     "styled-icons": "^10.33.0",
     "uuid": "^8.3.0",
