+++ peak-app/src/common/rich-text-editor/components.tsx
@@ -8,9 +8,7 @@ import {
     ELEMENT_H4,
     ELEMENT_H5,
     ELEMENT_H6,
-    ELEMENT_LI,
     ELEMENT_LINK,
-    ELEMENT_MEDIA_EMBED,
     ELEMENT_OL,
     ELEMENT_PARAGRAPH,
     ELEMENT_UL,
@@ -21,54 +19,40 @@ import {
 } from "@udecode/slate-plugins";
 import {DragIndicator} from "@styled-icons/material/DragIndicator";
 import React, {useMemo} from "react";
-import {
-    PEAK_LI_STYLE,
-    PEAK_OL_STYLE,
-    PEAK_UL_STYLE,
-    PeakBlockquoteElement, PeakCodeBlockElement
-} from "./plugins/slateComponentWrappers";
-import {TITLE} from "./types";
-import {PeakTitleElement} from "./plugins/peak-title-plugin/peak-title/PeakTitle";
 import {PEAK_CALLOUT} from "./plugins/peak-callout-plugin/defaults";
-import {PeakCalloutElement} from "./plugins/peak-callout-plugin/component/PeakCallout";
 import {PEAK_LEARNING} from "./plugins/peak-knowledge-plugin/constants";
 import {PeakLearningElement} from "./plugins/peak-knowledge-plugin/components/peak-learning-node/PeakLearningNode";
 import {PeakInlineLinkElement} from "./plugins/peak-link-plugin/inline-link/PeakHyperLink";
 import {
     ELEMENT_EMBED_STUB,
-    ELEMENT_TWITTER_EMBED,
-    ELEMENT_YOUTUBE_EMBED
 } from "./plugins/peak-media-embed-plugin/types";
 import {PeakMediaStubElement} from "./plugins/peak-media-embed-plugin/components/stub/MediaEmbedStub";
-import {
-    RichLinkEmbed,
-    TwitterEmbed,
-    YoutubeEmbed
-} from "./plugins/peak-media-embed-plugin/components/embedded_content/EmbeddedContent";
-import {ELEMENT_DIVIDER} from "./plugins/peak-divider";
-import {DividerElement} from "./plugins/peak-divider/element/DividerElement";
 import {clone} from "ramda";
 import {DRAGGABLE_ELEMENTS} from "./constants";
+import {PEAK_SLATE_COMPONENT_OVERRIDES} from "component-library/dist";
 
-const coreComponents = createSlatePluginsComponents({
-    [ELEMENT_BLOCKQUOTE]: PeakBlockquoteElement,
-})
+// const defaultComponents = createSlatePluginsComponents({
+//     [ELEMENT_BLOCKQUOTE]: PeakBlockquoteElement,
+//     [ELEMENT_LI]: PEAK_LI_STYLE,
+//     [ELEMENT_UL]: PEAK_UL_STYLE,
+//     [ELEMENT_OL]: PEAK_OL_STYLE,
+//     [ELEMENT_CODE_BLOCK]: PeakCodeBlockElement,
+//     [TITLE]: PeakTitleElement,
+//     [PEAK_CALLOUT]: PeakCalloutElement,
+//     [PEAK_LEARNING]: PeakLearningElement,
+//     [ELEMENT_LINK]: PeakInlineLinkElement,
+//     [ELEMENT_EMBED_STUB]: PeakMediaStubElement,
+//     [ELEMENT_TWITTER_EMBED]: TwitterEmbed,
+//     [ELEMENT_YOUTUBE_EMBED]: YoutubeEmbed,
+//     [ELEMENT_MEDIA_EMBED]: RichLinkEmbed,
+//     [ELEMENT_DIVIDER]: DividerElement,
+// })
 
 const defaultComponents = createSlatePluginsComponents({
-    [ELEMENT_BLOCKQUOTE]: PeakBlockquoteElement,
-    [ELEMENT_LI]: PEAK_LI_STYLE,
-    [ELEMENT_UL]: PEAK_UL_STYLE,
-    [ELEMENT_OL]: PEAK_OL_STYLE,
-    [ELEMENT_CODE_BLOCK]: PeakCodeBlockElement,
-    [TITLE]: PeakTitleElement,
-    [PEAK_CALLOUT]: PeakCalloutElement,
+    ...PEAK_SLATE_COMPONENT_OVERRIDES,
     [PEAK_LEARNING]: PeakLearningElement,
     [ELEMENT_LINK]: PeakInlineLinkElement,
-    [ELEMENT_EMBED_STUB]: PeakMediaStubElement,
-    [ELEMENT_TWITTER_EMBED]: TwitterEmbed,
-    [ELEMENT_YOUTUBE_EMBED]: YoutubeEmbed,
-    [ELEMENT_MEDIA_EMBED]: RichLinkEmbed,
-    [ELEMENT_DIVIDER]: DividerElement,
+    [ELEMENT_EMBED_STUB]: PeakMediaStubElement
 })
 
 const withStyledPlaceHolders = (components: any, placeholders: Options<PlaceholderProps>[]) => {
@@ -188,8 +172,3 @@ export const useComponents = (dnd: boolean, placeholders: Options<PlaceholderPro
     }, [dnd, placeholders])
 }
 
-export const useReadOnlyComponents = () => {
-    return useMemo(() => {
-        return clone(coreComponents)
-    }, [])
-}
