+++ peak-app/src/common/rich-text-editor/editorFactory.tsx
@@ -12,7 +12,7 @@ import {TNode} from "@udecode/slate-plugins-core/dist/types/TNode";
 import {useActiveEditorState} from "../../redux/slices/activeEditor/activeEditorSlice";
 import cn from "classnames"
 import "../../constants/peak-editor.scss"
-import {useComponents, useReadOnlyComponents} from "./components";
+import {useComponents} from "./components";
 import {DEFAULT_PLACEHOLDERS} from "./constants";
 import {contains, includes} from "ramda";
