--- peak-app/src/common/rich-text-editor/editors/journal/types.ts
@@ -1,5 +1,0 @@
-export interface JournalEntry {
-    id: string
-    entry_date: string,
-    body: any,
-}
