+++ peak-app/src/common/rich-text-editor/plugins/peak-knowledge-plugin/components/peak-knowledge-node/peak-tag-select/component/PeakTagSelect.tsx
@@ -55,6 +55,7 @@ export const PeakTagSelect = (props: { nodeId: number, nodePath: number[], selec
         const newTagList: PeakTag[] = displaySelectedTags.filter(tag => tag.title !== displayLabel.value as string)
         // User clicked on the X of the tag, without ever focusing
         if (!shouldFocus) {
+            // @ts-ignore
             Transforms.setNodes(editor, {selected_tags: newTagList}, { at: nodePath })
         }
         setSelectedTags(newTagList)
@@ -79,14 +80,18 @@ export const PeakTagSelect = (props: { nodeId: number, nodePath: number[], selec
     }
 
     const leaveDown = () => {
+        // @ts-ignore
         reEnterDown(editor, (n: Node) => isPeakKnowledgeNoteType(n) && n.id === nodeId)
     }
 
     // TODO: Why can't this be re-enter up?
     const leaveUp = () => {
+        // @ts-ignore
         const [theNode, path] = Editor.nodes(editor, { match: n => isPeakKnowledgeNoteType(n) && n.id === nodeId, at: []});
+        // @ts-ignore
         const [lastChildNode, wtf] = (theNode[0].children as Node[]).slice(-1)
 
+        // @ts-ignore
         if (lastChildNode.type === ELEMENT_CODE_BLOCK) {
             forceFocusToNode(lastChildNode)
         } else {
@@ -112,6 +117,7 @@ export const PeakTagSelect = (props: { nodeId: number, nodePath: number[], selec
             const newSelected: PeakTag[] = hotSwap(displaySelectedTags, createdTags)
             setTags([...tags, ...createdTags])
             setSelectedTags(newSelected)
+            // @ts-ignore
             Transforms.setNodes(editor, {selected_tags: newSelected}, { at: nodePath })
         }).finally(() => {
             setCurrentSearch("")
