+++ peak-app/src/common/rich-text-editor/plugins/peak-knowledge-plugin/components/peak-learning-node/PeakLearningNode.tsx
@@ -12,6 +12,7 @@ export const PeakLearningElement = (props: RenderElementProps) => {
     const { element } = props
     const editor = useTSlate()
     const path = ReactEditor.findPath(editor, props.element)
+    // @ts-ignore
     const tags = element.selected_tags as PeakTag[]
     const isEmpty: boolean = isNodeEmpty(element)
 
@@ -24,6 +25,7 @@ export const PeakLearningElement = (props: RenderElementProps) => {
             <div className="learning-body">
                 {props.children}
             </div>
+            {/* @ts-ignore */}
             <PeakTagSelect nodeId={element.id as number} nodePath={path} selected_tags={(tags) ? tags : []}/>
         </div>
     )
