+++ peak-app/src/common/rich-text-editor/plugins/peak-knowledge-plugin/utils.tsx
@@ -1,4 +1,3 @@
-import {useCallback, useEffect, useMemo, useState} from 'react';
 import {Editor, Node, Point, Range, Transforms} from "slate";
 import { previous } from "../../utils/base-utils";
 import { ReactEditor } from "slate-react";
@@ -9,10 +8,12 @@ import { ELEMENT_PARAGRAPH } from "@udecode/slate-plugins";
 import {UghEditorType} from "../../types";
 
 export const isNodeEmpty = (node: Node) => {
+    // @ts-ignore
     const theChildren: Node[] = node.children as Node[]
     if (!theChildren || theChildren.length != 1) { return false }
     const theNode = theChildren[0]
     const nodeText = Node.string(theNode)
+    // @ts-ignore
     return theNode.type === ELEMENT_PARAGRAPH && nodeText.length === 0
 }
 
@@ -21,6 +22,7 @@ export function isAtLastLineOfPeakKnowledgeNode(editor: Editor, nodeEntry?: any)
     const [currNode, currPath] = (nodeEntry) ? nodeEntry : Editor.above(editor)
     const [currParent, currParentPath] = Editor.parent(editor, currPath)
     const [lastChildNode] = currParent.children.slice(-1)
+    // @ts-ignore
     return isPeakKnowledgeNoteType(currParent) && lastChildNode.id === currNode.id
 }
 
@@ -46,6 +48,7 @@ export const knowledgeNodeOnKeyDownHandler: OnKeyDown = (editor: UghEditorType) 
         const [currParent, currParentPath] = Editor.parent(editor, currPath)
 
         const previousNode: Node | undefined = previous(editor)
+        // @ts-ignore
         if ((currParent && currParent.type !== ELEMENT_LI) && previousNode && isPeakKnowledgeNoteType(previousNode)) {
             console.log(`WE ARE DIRECTLY BELOW A KNOWLEDGE NODE`)
             event.preventDefault();
@@ -89,6 +92,7 @@ export const knowledgeNodeOnKeyDownHandler: OnKeyDown = (editor: UghEditorType) 
 // }
 
 export function isPeakKnowledgeNoteType(n: Node): boolean {
+    // @ts-ignore
     return PEAK_KNOWLEDGE_TYPES.includes(n.type as string)
 }
