+++ peak-app/src/common/rich-text-editor/plugins/peak-link-plugin/inline-link/PeakHyperLink.tsx
@@ -21,8 +21,11 @@ export const PeakInlineLinkElement = (props: RenderElementProps) => {
 
     const openUpMenu = (e: any) => {
         if (editorState.isEditing) {
+            // @ts-ignore
             const linkText: string = element.children[0].text as string
+            // @ts-ignore
             const linkUrl: string = element.url as string
+            // @ts-ignore
             const linkId: string = element.id as string
 
             const currentHyperlink: PeakHyperlinkState = {
@@ -33,6 +36,7 @@ export const PeakInlineLinkElement = (props: RenderElementProps) => {
             };
             dispatch(openEditLinkMenu({ hyperlinkState: currentHyperlink} ));
         } else {
+            // @ts-ignore
             const url: string = element.url as string
             if (url.startsWith("/")) {
                 e.preventDefault()
@@ -41,14 +45,17 @@ export const PeakInlineLinkElement = (props: RenderElementProps) => {
         }
     }
 
+    // @ts-ignore
     const externalLink: boolean = isExternalLink(element.url as string)
     return (
         <a
             {...attributes}
             data-slate-type={ELEMENT_LINK}
+            // @ts-ignore
             href={element.url as string}
             target={(externalLink) ? "_blank" : "_self"}
             className={"peak-hyperlink"}
+            // @ts-ignore
             key={element.link as string}
             onClick={(e) => openUpMenu(e)}>
                 {children}
