+++ peak-app/src/constants/editor.ts
@@ -1,4 +1,3 @@
-import {EMPTY_JOURNAL_STATE, JOURNAL_PAGE_ID} from "../common/rich-text-editor/editors/journal/constants";
 import {CHROME_EXTENSION} from "../common/rich-text-editor/editors/chrome-extension/constants";
 import {PeakEditorState, PeakHyperlinkState, PeakWikiPage} from "./wiki-types";
 import {EMPTY_PARAGRAPH_NODE} from "../common/rich-text-editor/editors/constants";
@@ -30,5 +29,4 @@ export const INITIAL_PAGE_STATE: PeakWikiPage = {
     title: ''
 };
 
-export const INITIAL_JOURNAL_STATE: PeakWikiPage = {...INITIAL_PAGE_STATE, body: EMPTY_JOURNAL_STATE, id: JOURNAL_PAGE_ID }
 export const INITIAL_CHROME_EXT_STATE: PeakWikiPage = {...INITIAL_PAGE_STATE, id: CHROME_EXTENSION }
