--- peak-app/src/views/journal/journal.scss
@@ -1,23 +1,0 @@
-@import "../../constants/variables";
-.peak-user-home {
-  display: flex;
-  flex-direction: column;
-
-  justify-content: flex-start;
-  align-items: flex-start;
-  min-height: 100%;
-  text-align: left;
-
-  &.empty {
-    align-items: center!important;
-    padding-top: 25%!important;
-  }
-
-  .peak-journal-title {
-    position: relative;
-    font-size: $title-font-size;
-    font-weight: $title-font-weight;
-    line-height: $title-line-height;
-    margin-bottom: 30px;
-  }
-}
